Blockchain self learning machine reinstalled
Blockchain is the only main chain that all other chains must  pass transaction through system to system before any interference of any human or by any member of any on chain or 2nd party. Or 3rd party and exc, all on chian or chains running transaction must be first  pass by verification  processing and  verify transaction value from both sides and then self verify by Blockchain self learning machine as it is the most secure way to confirm  funds and authentication  and cryptography  and SHA256++A. Code  Blockchain is the core and transaction operating system on the web 3 and all transaction must first be interceptor for all transactions with no adapters or rerouted or reroute to any one or any thing but only  blockchain self learning machine.